Refactor header into separate component
